allow OR search for regions
